Add functionality to process pending matches
The code updates include a new function, `process_to_be_played_matches`, for retrieving and processing matches that are yet to be played. We've also changed the functionality in the DAG file `grandmaster.py` to call the new `process_to_be_played_matches` function. Moreover, minor modifications in `retrieve_json_from_s3` and `generate_bracket_json` are carried out for a smoother workflow.